+++ CHANGES.rst
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 
+* Added support for :attr:`.User.city` to :class:`oauth2.LinkedIn` provider.
 * Fixed the API urls of the :class:`oauth1.Plurk` provider.
 * Added support for :attr:`.User.country` and :attr:`.User.city` to 
   :class: `oauth1.Twitter` provider.